chatbot_project/server/requirements.txt
ipywidgets==7.7.1
matplotlib==3.7.0
numpy==1.24.2
pandas==1.5.3
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
langchain==0.0.179
pinecone-client==2.2.1
fastapi
uvicorn
python-multipart
sqlalchemy
alembic
pytest